requirements.txt
gspread==5.11.3
python-telegram-bot==13.13
schedule==1.2.1